requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo;python_version<"3.9"
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.5
discord-webhook==1.0.0
eventkit==1.0.0
flake8==6.0.0
Flask==2.0.3
frozenlist==1.3.3
gunicorn==20.1.0
ib-insync==0.9.81
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.17.3
libtmux==0.21.0
loguru==0.6.0
MarkupSafe==2.1.1
mccabe==0.7.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
oauthlib==3.2.2
packaging==21.3
pandas==1.5.3
pycodestyle==2.10.0
pyflakes==3.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==13.6
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.3.5
regex==2021.4.4
requests==2.28.1
requests-oauthlib==1.3.1
shellingham==1.5.0
six==1.16.0
slack-webhook==1.0.7
toml==0.10.2
tornado==6.2
tweepy==4.12.1
typer==0.3.2
typer-cli==0.0.12
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.5
waitress==2.1.2
Werkzeug==2.0.3
yarl==1.8.2
